rename(controller): 패키지명 ui -> controller 변경 - Mapping annotation의 불필요한 속성 명세 제거
